3devo: Disable cin/cout
These seem to be included in the link even when unused (probably because
the constructor has side effects), so better just remove them
completely.

In FeFirmware, this saves just under 2k of flash and 264 bytes of RAM
(of which 80 bytes is the objects, and the rest seems to be a dozen or
so vtables).